Fix some whitespace and remove BigInteger usage

+++ src/main/java/com/tumblr/jumblr/JumblrClient.java
@@ -8,7 +8,6 @@ import com.tumblr.jumblr.responses.ResponseWrapper;
 import com.tumblr.jumblr.types.Blog;
 import com.tumblr.jumblr.types.Post;
 import com.tumblr.jumblr.types.User;
-import java.math.BigInteger;
 import java.net.HttpURLConnection;
 import java.util.HashMap;
 import java.util.List;
@@ -37,9 +36,9 @@ public final class JumblrClient {
         this.service = new ServiceBuilder().
             provider(TumblrApi.class).
             apiKey(consumerKey).apiSecret(consumerSecret).
-            build();        
+            build();
     }
-    
+
     /**
      * Instantiate a new Jumblr Client
      * @param consumerKey The consumer key for the client
@@ -60,7 +59,7 @@ public final class JumblrClient {
     public void setToken(String token, String tokenSecret) {
         this.token = new Token(token, tokenSecret);
     }
-    
+
     /**
      * Get the user info for the authenticated User
      * @return The authenticated user
@@ -68,7 +67,7 @@ public final class JumblrClient {
     public User user() {
         return this.clearGet("/user/info").getUser();
     }
-    
+
     /**
      * Get the user dashboard for the authenticated User
      * @param options the options for the call (or null)
@@ -77,11 +76,11 @@ public final class JumblrClient {
     public List<Post> userDashboard(Map<String, ?> options) {
         return this.clearGet("/user/dashboard", options).getPosts();
     }
-    
+
     public List<Post> userDashboard() {
         return this.userDashboard(null);
     }
-    
+
     /**
      * Get the blogs the given user is following
      * @return a List of blogs
@@ -91,7 +90,7 @@ public final class JumblrClient {
     }
 
     public List<Blog> userFollowing() { return this.userFollowing(null); }
-    
+
     /**
      * Get the blog info for a given blog
      * @param blogName the Name of the blog
@@ -102,7 +101,7 @@ public final class JumblrClient {
         map.put("api_key", this.apiKey);
         return this.clearGet(JumblrClient.blogPath(blogName, "/info"), map).getBlog();
     }
-    
+
     /**
      * Get the followers for a given blog
      * @param blogName the name of the blog
@@ -113,7 +112,7 @@ public final class JumblrClient {
     }
 
     public List<User> blogFollowers(String blogName) { return this.blogFollowers(blogName, null); }
-    
+
     /**
      * Get the public likes for a given blog
      * @param blogName the name of the blog
@@ -128,11 +127,11 @@ public final class JumblrClient {
         soptions.put("api_key", this.apiKey);
         return this.clearGet(JumblrClient.blogPath(blogName, "/likes"), options).getLikedPosts();
     }
-    
+
     public List<Post> blogLikes(String blogName) {
         return this.blogLikes(blogName, null);
     }
-    
+
     /**
      * Get the posts for a given blog
      * @param blogName the name of the blog
@@ -153,11 +152,11 @@ public final class JumblrClient {
         }
         return this.clearGet(JumblrClient.blogPath(blogName, path), options).getPosts();
     }
-    
+
     public List<Post> blogPosts(String blogName) {
         return this.blogPosts(blogName, null);
     }
-    
+
     /**
      * Get the queued posts for a given blog
      * @param blogName the name of the blog
@@ -167,11 +166,11 @@ public final class JumblrClient {
     public List<Post> blogQueuedPosts(String blogName, Map<String, ?> options) {
         return this.clearGet(JumblrClient.blogPath(blogName, "/posts/queue"), options).getPosts();
     }
-    
+
     public List<Post> blogQueuedPosts(String blogName) {
         return this.blogQueuedPosts(blogName, null);
     }
-    
+
     /**
      * Get the draft posts for a given blog
      * @param blogName the name of the blog
@@ -181,7 +180,7 @@ public final class JumblrClient {
     public List<Post> blogDraftPosts(String blogName, Map<String, ?> options) {
         return this.clearGet(JumblrClient.blogPath(blogName, "/posts/draft"), options).getPosts();
     }
-    
+
     public List<Post> blogDraftPosts(String blogName) {
         return this.blogDraftPosts(blogName, null);
     }
@@ -195,11 +194,11 @@ public final class JumblrClient {
     public List<Post> blogSubmissions(String blogName, Map<String, ?> options) {
         return this.clearGet(JumblrClient.blogPath(blogName, "/posts/submission"), options).getPosts();
     }
-    
+
     public List<Post> blogSubmissions(String blogName) {
         return this.blogSubmissions(blogName, null);
-    }    
-    
+    }
+
     /**
      * Get the likes for the authenticated user
      * @param options the options for this call (or null)
@@ -208,11 +207,11 @@ public final class JumblrClient {
     public List<Post> userLikes(Map<String, ?> options) {
         return this.clearGet("/user/likes", options).getLikedPosts();
     }
-    
+
     public List<Post> userLikes() {
         return this.userLikes(null);
     }
-    
+
     /**
      * Get a specific size avatar for a given blog
      * @param blogName the avatar URL of the blog
@@ -229,35 +228,35 @@ public final class JumblrClient {
             return response.getHeader("Location");
         } else {
             throw new JumblrException(response);
-        }        
+        }
     }
 
     public String blogAvatar(String blogName) { return this.blogAvatar(blogName, null); }
- 
+
     /**
      * Like a given post
      * @param postId the ID of the post to like
      * @param reblogKey The reblog key for the post
      */
-    public void like(BigInteger postId, String reblogKey) {
+    public void like(Long postId, String reblogKey) {
         Map map = new HashMap<String, String>();
         map.put("id", postId.toString());
         map.put("reblog_key", reblogKey);
         this.clearPost("/user/like", map);
     }
-    
+
     /**
      * Unlike a given post
      * @param postId the ID of the post to unlike
      * @param reblogKey The reblog key for the post
      */
-    public void unlike(BigInteger postId, String reblogKey) {
+    public void unlike(Long postId, String reblogKey) {
         Map map = new HashMap<String, String>();
         map.put("id", postId.toString());
         map.put("reblog_key", reblogKey);
         this.clearPost("/user/unlike", map);
     }
-    
+
     /**
      * Follow a given blog
      * @param blogName The name of the blog to follow
@@ -267,7 +266,7 @@ public final class JumblrClient {
         map.put("url", JumblrClient.blogUrl(blogName));
         this.clearPost("/user/follow", map);
     }
-    
+
     /**
      * Unfollow a given blog
      * @param blogName the name of the blog to unfollow
@@ -277,18 +276,18 @@ public final class JumblrClient {
         map.put("url", JumblrClient.blogUrl(blogName));
         this.clearPost("/user/follow", map);
     }
-    
+
     /**
      * Delete a given post
      * @param blogName the name of the blog the post is in
      * @param postId the id of the post to delete
      */
-    public void postDelete(String blogName, BigInteger postId) {
+    public void postDelete(String blogName, Long postId) {
         Map map = new HashMap<String, String>();
         map.put("id", postId);
         this.clearPost(JumblrClient.blogPath(blogName, "/post/delete"), map);
     }
-    
+
     /**
      * Reblog a given post
      * @param blogName the name of the blog the post is in
@@ -296,7 +295,7 @@ public final class JumblrClient {
      * @param reblogKey the reblog_key of the post
      * @param options Additional options (or null)
      */
-    public Post postReblog(String blogName, BigInteger postId, String reblogKey, Map<String, ?> options) {
+    public Post postReblog(String blogName, Long postId, String reblogKey, Map<String, ?> options) {
         if (options == null) {
             options = new HashMap<String, String>();
         }
@@ -305,30 +304,30 @@ public final class JumblrClient {
         soptions.put("reblog_key", reblogKey);
         return this.clearPost(JumblrClient.blogPath(blogName, "/post/reblog"), options).getPost();
     }
-    
-    public Post postReblog(String blogName, BigInteger postId, String reblogKey) {
+
+    public Post postReblog(String blogName, Long postId, String reblogKey) {
         return this.postReblog(blogName, postId, reblogKey, null);
     }
-    
+
     /**
      **
      **
      */
-    
+
     private ResponseWrapper clearGet(String path) {
         return this.clearGet(path, null);
     }
-    
+
     private ResponseWrapper clearPost(String path, Map<String, ?> bodyMap) {
         Response response = this.post(path, bodyMap);
         return this.clear(response);
     }
-    
+
     private ResponseWrapper clearGet(String path, Map<String, ?> map) {
         Response response = this.get(path, map);
         return this.clear(response);
     }
-    
+
     private ResponseWrapper clear(Response response) {
         if (response.getCode() == 200 || response.getCode() == 201) {
             String json = response.getBody();
@@ -344,11 +343,11 @@ public final class JumblrClient {
             throw new JumblrException(response);
         }
     }
-    
+
     private Response get(String path) {
         return this.get(path, null);
     }
-    
+
     private Response get(String path, Map<String, ?> queryParams) {
         String url = "http://api.tumblr.com/v2" + path;
         OAuthRequest request = new OAuthRequest(Verb.GET, url);
@@ -372,7 +371,7 @@ public final class JumblrClient {
         service.signRequest(token, request);
         return request.send();
     }
-    
+
     private static String blogPath(String blogName, String extPath) {
         String bn = blogName.contains(".") ? blogName : blogName + ".tumblr.com";
         return "/blog/" + bn + extPath;
